+++ src/survey.ts
@@ -49,12 +49,12 @@ import { CssClassBuilder } from "./utils/cssClassBuilder";
  */
 export class SurveyModel extends Base
   implements
-    ISurvey,
-    ISurveyData,
-    ISurveyImpl,
-    ISurveyTriggerOwner,
-    ISurveyErrorOwner,
-    ILocalizableOwner {
+  ISurvey,
+  ISurveyData,
+  ISurveyImpl,
+  ISurveyTriggerOwner,
+  ISurveyErrorOwner,
+  ILocalizableOwner {
   public static readonly TemplateRendererComponentName: string =
     "sv-template-renderer";
   [index: string]: any;
@@ -1194,7 +1194,7 @@ export class SurveyModel extends Base
    * Gets or sets whether the first input is focused on showing a next or a previous page.
    */
   public get focusFirstQuestionAutomatic(): boolean {
-    return this.getPropertyValue("focusFirstQuestionAutomatic", true);
+    return this.getPropertyValue("focusFirstQuestionAutomatic");
   }
   public set focusFirstQuestionAutomatic(val: boolean) {
     this.setPropertyValue("focusFirstQuestionAutomatic", val);
@@ -1204,7 +1204,7 @@ export class SurveyModel extends Base
    * Set this property to `false` (the default value is `true`) if you do not want to bring the focus to the first question that has error on the page.
    */
   public get focusOnFirstError(): boolean {
-    return this.getPropertyValue("focusOnFirstError", true);
+    return this.getPropertyValue("focusOnFirstError");
   }
   public set focusOnFirstError(val: boolean) {
     this.setPropertyValue("focusOnFirstError", val);
@@ -1233,7 +1233,7 @@ export class SurveyModel extends Base
    * @see showNavigationButtons
    */
   public get showPrevButton(): boolean {
-    return this.getPropertyValue("showPrevButton", true);
+    return this.getPropertyValue("showPrevButton");
   }
   public set showPrevButton(val: boolean) {
     this.setPropertyValue("showPrevButton", val);
@@ -1243,7 +1243,7 @@ export class SurveyModel extends Base
    * @see title
    */
   public get showTitle(): boolean {
-    return this.getPropertyValue("showTitle", true);
+    return this.getPropertyValue("showTitle");
   }
   public set showTitle(val: boolean) {
     this.setPropertyValue("showTitle", val);
@@ -1253,7 +1253,7 @@ export class SurveyModel extends Base
    * @see PageModel.title
    */
   public get showPageTitles(): boolean {
-    return this.getPropertyValue("showPageTitles", true);
+    return this.getPropertyValue("showPageTitles");
   }
   public set showPageTitles(val: boolean) {
     this.setPropertyValue("showPageTitles", val);
@@ -1265,7 +1265,7 @@ export class SurveyModel extends Base
    * @see navigateToUrl
    */
   public get showCompletedPage(): boolean {
-    return this.getPropertyValue("showCompletedPage", true);
+    return this.getPropertyValue("showCompletedPage");
   }
   public set showCompletedPage(val: boolean) {
     this.setPropertyValue("showCompletedPage", val);
@@ -1366,7 +1366,7 @@ export class SurveyModel extends Base
    * @see commentPrefix
    */
   public get storeOthersAsComment(): boolean {
-    return this.getPropertyValue("storeOthersAsComment", true);
+    return this.getPropertyValue("storeOthersAsComment");
   }
   public set storeOthersAsComment(val: boolean) {
     this.setPropertyValue("storeOthersAsComment", val);
@@ -1378,7 +1378,7 @@ export class SurveyModel extends Base
    * @see maxOthersLength
    */
   public get maxTextLength(): number {
-    return this.getPropertyValue("maxTextLength", 0);
+    return this.getPropertyValue("maxTextLength");
   }
   public set maxTextLength(val: number) {
     this.setPropertyValue("maxTextLength", val);
@@ -1392,7 +1392,7 @@ export class SurveyModel extends Base
    * @see maxTextLength
    */
   public get maxOthersLength(): number {
-    return this.getPropertyValue("maxOthersLength", 0);
+    return this.getPropertyValue("maxOthersLength");
   }
   public set maxOthersLength(val: number) {
     this.setPropertyValue("maxOthersLength", val);
@@ -1438,6 +1438,15 @@ export class SurveyModel extends Base
   }
   public set checkErrorsMode(val: string) {
     this.setPropertyValue("checkErrorsMode", val);
+  }
+  /**
+   * Gets or sets a value indicating whether comment questions are automatically increase height to display its entire contents.
+   */
+  public get autoGrowComment(): boolean {
+    return this.getPropertyValue("autoGrowComment");
+  }
+  public set autoGrowComment(val: boolean) {
+    this.setPropertyValue("autoGrowComment", val);
   }
   /**
    * Gets or sets a value that specifies how the survey updates its questions' text values.
@@ -2784,7 +2793,7 @@ export class SurveyModel extends Base
         return info.requiredQuestionCount > 1
           ? Math.ceil(
             (info.requiredAnsweredQuestionCount * 100) /
-                info.requiredQuestionCount
+            info.requiredQuestionCount
           )
           : 100;
       }
@@ -3346,7 +3355,7 @@ export class SurveyModel extends Base
    * - `questionPerPage` - show one question per page. Survey will create a separate page for every question.
    */
   public get questionsOnPageMode(): string {
-    return this.getPropertyValue("questionsOnPageMode", "standard");
+    return this.getPropertyValue("questionsOnPageMode");
   }
   public set questionsOnPageMode(val: string) {
     this.setPropertyValue("questionsOnPageMode", val);
@@ -3378,7 +3387,7 @@ export class SurveyModel extends Base
    * @see editText
    */
   public get showPreviewBeforeComplete(): string {
-    return this.getPropertyValue("showPreviewBeforeComplete", "noPreview");
+    return this.getPropertyValue("showPreviewBeforeComplete");
   }
   public set showPreviewBeforeComplete(val: string) {
     this.setPropertyValue("showPreviewBeforeComplete", val);
@@ -3571,18 +3580,18 @@ export class SurveyModel extends Base
     var savingDataStarted = false;
     var onCompleteOptions = {
       isCompleteOnTrigger: isCompleteOnTrigger,
-      showDataSaving: function(text: string) {
+      showDataSaving: function (text: string) {
         savingDataStarted = true;
         self.setCompletedState("saving", text);
       },
-      showDataSavingError: function(text: string) {
+      showDataSavingError: function (text: string) {
         self.setCompletedState("error", text);
       },
-      showDataSavingSuccess: function(text: string) {
+      showDataSavingSuccess: function (text: string) {
         self.setCompletedState("success", text);
         self.navigateTo();
       },
-      showDataSavingClear: function(text: string) {
+      showDataSavingClear: function (text: string) {
         self.setCompletedState("", "");
       },
     };
@@ -3622,7 +3631,7 @@ export class SurveyModel extends Base
     this.setPropertyValue("isValidatingOnServer", val);
     this.onIsValidatingOnServerChanged();
   }
-  protected onIsValidatingOnServerChanged() {}
+  protected onIsValidatingOnServerChanged() { }
   protected doServerValidation(
     doComplete: boolean,
     isPreview: boolean = false
@@ -3638,7 +3647,7 @@ export class SurveyModel extends Base
       data: <{ [index: string]: any }>{},
       errors: {},
       survey: this,
-      complete: function() {
+      complete: function () {
         self.completeServerValidation(options, isPreview);
       },
     };
@@ -4641,7 +4650,7 @@ export class SurveyModel extends Base
     if (
       this.isValueChangedOnRunningCondition &&
       this.conditionRunnerCounter <
-        settings.maximumConditionRunCountOnValueChanged
+      settings.maximumConditionRunCountOnValueChanged
     ) {
       this.isValueChangedOnRunningCondition = false;
       this.conditionRunnerCounter++;
@@ -4726,7 +4735,7 @@ export class SurveyModel extends Base
     this.createSurveyService().sendResult(
       postId,
       this.data,
-      function(success: boolean, response: any, request: any) {
+      function (success: boolean, response: any, request: any) {
         if (self.surveyShowDataSaving) {
           if (success) {
             self.setCompletedState("success", "");
@@ -4752,7 +4761,7 @@ export class SurveyModel extends Base
    */
   public getResult(resultId: string, name: string) {
     var self = this;
-    this.createSurveyService().getResult(resultId, name, function(
+    this.createSurveyService().getResult(resultId, name, function (
       success: boolean,
       data: any,
       dataList: any[],
@@ -4791,7 +4800,7 @@ export class SurveyModel extends Base
       this.createSurveyService().getSurveyJsonAndIsCompleted(
         this.surveyId,
         this.clientId,
-        function(
+        function (
           success: boolean,
           json: string,
           isCompleted: string,
@@ -4805,7 +4814,7 @@ export class SurveyModel extends Base
         }
       );
     } else {
-      this.createSurveyService().loadSurvey(this.surveyId, function(
+      this.createSurveyService().loadSurvey(this.surveyId, function (
         success: boolean,
         result: string,
         response: any
@@ -4824,8 +4833,8 @@ export class SurveyModel extends Base
     this.onLoadSurveyFromService();
     this.onLoadedSurveyFromService.fire(this, {});
   }
-  protected onLoadingSurveyFromService() {}
-  protected onLoadSurveyFromService() {}
+  protected onLoadingSurveyFromService() { }
+  protected onLoadSurveyFromService() { }
   private resetVisibleIndexes() {
     var questions = this.getAllQuestions(true);
     for (var i = 0; i < questions.length; i++) {
@@ -4896,8 +4905,8 @@ export class SurveyModel extends Base
     this.isEndLoadingFromJson = null;
     this.updateVisibleIndexes();
   }
-  protected onBeforeCreating() {}
-  protected onCreating() {}
+  protected onBeforeCreating() { }
+  protected onCreating() { }
   private getProcessedTextValue(textValue: TextPreProcessorValue): void {
     this.getProcessedTextValueCore(textValue);
     if (!this.onProcessTextValue.isEmpty) {
@@ -5602,7 +5611,7 @@ export class SurveyModel extends Base
    * @see stopTimer
    */
   public get showTimerPanel(): string {
-    return this.getPropertyValue("showTimerPanel", "none");
+    return this.getPropertyValue("showTimerPanel");
   }
   public set showTimerPanel(val: string) {
     this.setPropertyValue("showTimerPanel", val);
@@ -5626,7 +5635,7 @@ export class SurveyModel extends Base
    * @see onTimerPanelInfoText
    */
   public get showTimerPanelMode(): string {
-    return this.getPropertyValue("showTimerPanelMode", "all");
+    return this.getPropertyValue("showTimerPanelMode");
   }
   public set showTimerPanelMode(val: string) {
     this.setPropertyValue("showTimerPanelMode", val);
@@ -5714,7 +5723,7 @@ export class SurveyModel extends Base
   public startTimer() {
     if (this.isTimerStarted || this.isDesignMode) return;
     var self = this;
-    this.timerFunc = function() {
+    this.timerFunc = function () {
       self.doTimer();
     };
     this.isTimerStarted = true;
@@ -5952,10 +5961,10 @@ Serializer.addClass("survey", [
     baseClassName: "question",
     visible: false,
     isLightSerializable: false,
-    onGetValue: function(obj: any): any {
+    onGetValue: function (obj: any): any {
       return null;
     },
-    onSetValue: function(obj: any, value: any, jsonConverter: any) {
+    onSetValue: function (obj: any, value: any, jsonConverter: any) {
       obj.pages.splice(0, obj.pages.length);
       var page = obj.addNewPage("");
       jsonConverter.toObject({ questions: value }, page);
@@ -6047,6 +6056,7 @@ Serializer.addClass("survey", [
     default: "onBlur",
     choices: ["onBlur", "onTyping"],
   },
+  { name: "autoGrowComment:boolean", default: false },
   { name: "startSurveyText", serializationProperty: "locStartSurveyText" },
   { name: "pagePrevText", serializationProperty: "locPagePrevText" },
   { name: "pageNextText", serializationProperty: "locPageNextText" },
@@ -6057,7 +6067,7 @@ Serializer.addClass("survey", [
   {
     name: "questionStartIndex",
     dependsOn: ["showQuestionNumbers"],
-    visibleIf: function(survey: any) {
+    visibleIf: function (survey: any) {
       return !survey || survey.showQuestionNumbers !== "off";
     },
   },