+++ src/vue/comment.vue
@@ -12,16 +12,13 @@
       v-bind:aria-label="question.locTitle.renderedHtml"
       :rows="question.rows"
       :placeholder="question.renderedPlaceHolder"
-      :class="
-        question.cssClasses ? question.cssClasses.root : 'panel-comment-root'
-      "
+      :class="question.cssClasses ? question.cssClasses.root : 'panel-comment-root'"
       @change="change"
-      @keyup="keyup"
+      @input="onInput"
       v-bind:aria-required="question.isRequired"
       :aria-invalid="question.errors.length > 0"
-      :aria-describedby="
-        question.errors.length > 0 ? question.id + '_errors' : null
-      "
+      :aria-describedby="question.errors.length > 0 ? question.id + '_errors' : null"
+      v-bind:style="{ resize: question.autoGrow ? 'none' : 'both' }"
     ></textarea>
     <div v-if="question.isReadOnlyRenderDiv()">{{ question.value }}</div>
   </div>
@@ -38,9 +35,9 @@ export class Comment extends QuestionVue<QuestionCommentModel> {
   change(event: any) {
     this.question.value = event.target.value;
   }
-  keyup(event: any) {
-    if (!this.question.isInputTextUpdate) return;
-    this.question.value = event.target.value;
+  onInput(event: any) {
+    if (this.question.isInputTextUpdate) this.question.value = event.target.value;
+    else this.question.updateElement();
   }
 }
 Vue.component("survey-comment", Comment);